Thimoo Challenge 2017
## The challenge
The goal is to create an application that can manage clients' offers for thimoo. To do this,
you have to use the REST API described [here](https://github.com/thimoo/challenge-api).
This API offers four endpoints to retrieve and save the content. Your application should
look something like this:
The offer has a form title, a client name and is composed of a list of products selectable in
a list at the right of the offer. A product can only be selectable one time (to simplify the API)
in the offer and has a quantity attached. When we click on a product, if it is not already
present in the offer, he is added. If you have any other ideas to improve this concept, feel
free!